+++ R/CameraRenaming.R
@@ -260,29 +260,78 @@ cameraRename2 <- function(in.dir, out.dir=NULL, file.type, trigger.info=NULL, re
 }
 
 ## Another major update to the renaming function (Added 2022-08-25) ####
-##' @description This function is used to extract metadata information and rename camera trap images using date-times and an assigned serial number to each image. This function has all the same capabilities as the cameraRename2 function but should handle large datasets and corrupt files better. Unlike the cameraRename or cameraRename2 functions, this version keeps everything within each camera directory, running the same process on each camera. This allows it to take advantage of parallel processing and other functions to improve speed for very large datasets.
+##' @description This function is used to extract metadata information and rename camera trap images using date-times and an assigned serial number to each image.
+##' This function has all the same capabilities as the cameraRename2 function but should handle large datasets and corrupt files better.
+##' Unlike the cameraRename2 functions, this version keeps everything within each camera directory, running the same process on each camera.
+##' This allows it to take advantage of parallel processing and other functions to improve speed for very large datasets.
 ##'
 ##' @title Rename Camera Trap Images
 ##'
-##' @param in.dir String. The directory where the pictures are located. I typically specify a folder containing the camera folders although any folder containing images somewhere in the directory tree could work. The function can look for either ".jpg" or ".mp4" files. See file.type.
-##' @param out.dir String or Vector. The directory where the pictures will be renamed to. The default is NULL which will use the in.dir as the location for images. Another way to specify the output location to be the input location is to set out.dir = "in.dir". You can use this to change the output location for each camera directory. You can specify this with a single string, in which case it will use the relative paths to the images to locate the directory location or you can specify a vector of directories. The length of this vector must equal the number of camera directories in the in.dir. If rename="copy", you should specify a different out.dir than in.dir although the function does not check for this.
-##' @param file.type String. The type of file you want to rename. For images, this is likely to be ".jpg". For videos, this is likely to be ".MP4". This can also take c("image", "video"). For "image", this will default the file type to ".jpg", for "video", it will default to ".mp4". If a different file type is needed, manually specify the type. You must include the "." in the file type. This function can theoretically take any image format, although I do not know how exiftool reads file types other than ".jpg" and ".mp4". The function will ignore case when looking for file types.
-##' @param trigger.info String or NULL. Should additional information besides the date-time be included in the output? This defaults to NULL where no additional information is included. Because camera-specific information is variable between camera models, you must specify the camera model if you want additional information. Currently, only c("Hyperfire2", "Ultrafire_Video") are supported. Note that Ultrafire pictures use the same metadata tags as the Hyperfire2. Additional camera models could be added. See details for more information on this.
-##' @param rename String. The default is none. Other options include c("replace", "copy"). This specifies how images should be renamed. "none" tells R to not rename any pictures, "replace" replaces the name then moves the image with its new name to the out.dir, and "copy" creates a new copy of the image with its new name in the out.dir. Copy has not been tested with an out.dir the same as the in.dir but it is likely that it will keep both images and you will end up with duplicates. I suggest choosing a different out.dir if you are going to copy. Copy also takes significantly longer than replace.
-##' @param return.type String. How should the data be outputted. Options are c("list", "df"). The default is as a list where each camera directory is in its own item in the list. If you output a "df", the list of each camera is converted into a single data.frame object.
-##' @param adjust String or NULL. Do the image date-times need adjustment? This defaults to NULL, indicating no adjustment needed. You only need to specify this if your image date-times need to be adjusted. This could arise due to daylight savings time or misentered date-time information on the camera. This can be thought of as an R version of SpecialRenamer. You can either specify a difftime object or a character vector of the original and new date-times. If you specify a vector or original and new date-times, this must have length==2.
-##' @param fix.names Logical. Should a serial number be appended to date-time names? The default is FALSE. This was added to allow for updating image names to include a unique serial number that have already been run through a program like Renamer or SpecialRenamer where it may be uncertain if the DateTimeOriginal in the Exif data is accurate or not. Only set this to TRUE if you know what you are doing. It is highly recommended that you set rename="none" for this then rename manually after running the function.
-##' @param pp Logical. Should this function take advantage of parallel processing. The default is FALSE. Because the function separates tasks by camera directory, it can use parallel processing to run multiple cameras at the same time. This is currently set up to run in Windows OS so I do not know if this will work on a Mac or Linux system. If you want this functionality on a Unix device and know how to set it up, let me know and I can incorporate it.
-##' @param cores.left Numeric. How many cores should be left available when using parallel processing? The default is NULL. This is only necessary when pp=TRUE. If left at the default, the function will default to 2 cores remaining which is generally enough to continue using a PC while the function runs. I would set this to be greater than 0, otherwise the function will use the entire processing power of your computer. To see how many cores you have available for parallel processing, use: parallel::detectCores().
+##' @param in.dir String. The directory where the pictures are located.
+##' I typically specify a folder containing the camera folders although any folder containing images somewhere in the directory tree could work.
+##' The function can look for either ".jpg" or ".mp4" files. See file.type.
+##' @param out.dir String or Vector. The directory where the pictures will be renamed to.
+##' The default is NULL which will use the in.dir as the location for images.
+##' Another way to specify the output location to be the input location is to set out.dir = "in.dir".
+##' You can use this to change the output location for each camera directory.
+##' You can specify this with a single string, in which case it will use the relative paths to the images to locate the directory location or you can specify a vector of directories.
+##' The length of this vector must equal the number of camera directories in the in.dir. If rename="copy", you should specify a different out.dir than in.dir although the function does not check for this.
+##' @param file.type String. The type of file you want to rename.
+##' For images, this is likely to be ".jpg". For videos, this is likely to be ".mp4".
+##' This can also take c("image", "video"). For "image", this will default the file type to ".jpg", for "video", it will default to ".mp4".
+##' If a different file type is needed, this function will need to be updated to accommodate this. Please let me know and I will add it.
+##' This function can theoretically take any image format, although I do not know how exiftool reads file types other than ".jpg" and ".mp4". The function will ignore case when looking for file types.
+##' @param trigger.info String or NULL. Should additional information besides the date-time be included in the output?
+##' This defaults to NULL where no additional information is included, only date-time information.
+##' Because camera-specific information is variable between camera models, you must specify the camera model if you want additional information.
+##' Currently, only c("Hyperfire2", "Ultrafire_Video") are supported. Note that Reconyx Ultrafire pictures use the same metadata tags as the Reconyx Hyperfire2.
+##' Additional camera models could be added. See details for more information on this.
+##' @param rename String. The default is none. Other options include c("replace", "copy").
+##' This specifies how images should be renamed.
+##' "none" tells R to not rename any pictures, "replace" replaces the name then moves the image with its new name to the out.dir, and "copy" creates a new copy of the image with its new name in the out.dir.
+##' Copy has not been tested with an out.dir the same as the in.dir but it is likely that it will keep both images and you will end up with duplicates.
+##' I suggest choosing a different out.dir if you are going to copy. Copy also takes significantly longer than replace.
+##' @param return.type String. How should the data be outputted.
+##' Options are c("list", "df"). The default is as a list where each camera directory is in its own item in the list.
+##' If you output a "df", the list of each camera is converted into a single data.frame object containing all cameras.
+##' @param adjust String or NULL. Do the image date-times need adjustment?
+##' This defaults to NULL, indicating no adjustment needed.
+##' You only need to specify this if your image date-times need to be adjusted.
+##' This could arise due to daylight savings time or misentered date-time information on the camera.
+##' This can be thought of as an R version of SpecialRenamer.
+##' You can either specify a difftime object or a character vector of the original and new date-times.
+##' If you specify a vector or original and new date-times, this must have length==2.
+##' @param fix.names Logical. Should a serial number be appended to date-time names?
+##' The default is FALSE. This was added to allow for updating image names to include a unique serial number that have already been run through a program like Renamer or SpecialRenamer where it may be uncertain if the DateTimeOriginal in the Exif data is accurate or not.
+##' Only set this to TRUE if you know what you are doing. It is highly recommended that you set rename="none" for this then rename manually after running the function.
+##' @param pp Logical. Should this function take advantage of parallel processing.
+##' The default is FALSE. Because the function separates tasks by camera directory, it can use parallel processing to run multiple cameras at the same time.
+##' This is currently set up to run in Windows OS so I do not know if this will work on a Mac or Linux system.
+##' If you want this functionality on a Unix device and know how to set it up, let me know and I can incorporate it.
+##' @param cores.left Numeric. How many cores should be left available when using parallel processing?
+##' The default is NULL. This is only necessary when pp=TRUE.
+##' If left at the default, the function will default to 2 cores remaining which is generally enough to continue using a PC while the function runs.
+##' I would set this to be greater than 0, otherwise the function will use the entire processing power of your computer.
+##' To see how many cores you have available for parallel processing, use: parallel::detectCores().
 ##'
-##' @details Important: Exiftool must be available on the machine in its default location for the function to work. This function will check for exiftool and install it into the default directory if it is not available before running. Previous versions of this function required you to manually check this but this requirement has been removed.
+##' @details Important: Exiftool must be available on the machine in its default location for the function to work.
+##' This function will check for exiftool and install it into the default directory if it is not available before running.
+##' Previous versions of this function required you to manually check this but this requirement has been removed.
 ##'
-##' If you are interested in additional metadata information that is not provided, please run exiftoolr::exif_read(["at least one image"]) and choose the columns you are interested in. Send me the exact column names, and your camera model and I can add an option to trigger.info. Please keep in mind that the more metadata tags you choose, the longer the function takes to run.
+##' When running this function on video files, you need to make sure you specify a trigger.info that accommodates videos.
+##' Video files doe not have a DateTimeOriginal field in their metadata so it will fail to run if you do not properly specify the correct field.
+##' Right now, the "Ultrafire_video" option only accesses the "CreateDate" field which is used for video files.
 ##'
-##' If you need to fix names (or add a serial number to existing names) rather than replace names, then you should specify fix.names to TRUE. By specifying this, it is telling the tool to check if the images are in the form YYYY MM DD HH MM SS.jpg (or, more likely some consistent form with spaces but I don't know for sure). When it does this, if an image does not fit this form, it replaces the name with the new date-time information derived from the metadata. The reason for this is that sometimes camera date-times are wrong so we needed to make an adjustment using SpecialRename, similar to what is done with the adjust call in this function. Because I did not want to go back through the images and find those that were orignally fixed and fix them again, this should leave those photos alone and only adjust improperly named images.
+##' If you are interested in additional metadata information that is not provided, please run exiftoolr::exif_read(["at least one image"]) and choose the columns you are interested in.
+##' Send me the exact column names, and your camera model and I can add an option to trigger.info. Please keep in mind that the more metadata tags you choose, the longer the function takes to run.
+##'
+##' If you need to fix names (or add a serial number to existing names) rather than replace names, then you should specify fix.names to TRUE.
+##' By specifying this, it is telling the tool to check if the images are in the form YYYY MM DD HH MM SS.jpg (or, more likely some consistent form with spaces but I don't know for sure).
+##' When it does this, if an image does not fit this form, it replaces the name with the new date-time information derived from the metadata.
+##' The reason for this is that sometimes camera date-times are wrong so we needed to make an adjustment using SpecialRename, similar to what is done with the adjust call in this function.
+##' Because I did not want to go back through the images and find those that were orignally fixed and fix them again, this should leave those photos alone and only adjust improperly named images.
 ##'
 ##' @return This function outputs either a list or a data.frame, depending on whether return.type = c("list", "df")
-##'
 ##' @return list:
 ##' If a list, each camera directory will be kept in a separate item in the list. This allows for easy checking if there is a problem which would have thrown an error when running exiftool, namely from a corrupt file. After outputting a list, you can combine the items into a single data.frame using do.call(rbind, out) where out is the name of the list object outputted by this function.
 ##' @return df:
@@ -332,16 +381,43 @@ cameraRename3 <- function(in.dir, out.dir=NULL, file.type, trigger.info=NULL, re
   #cores.left <- 2
 
   print(paste("This function started at ", Sys.time(), ". Loading images...", sep = ""))
-  # Define which metadata tags you want to include in your output
-  if(isTRUE(is.null(trigger.info))){
-    Tag <- c("DateTimeOriginal")
-  }else if(trigger.info=="Hyperfire2"){
-    Tag <- c("DateTimeOriginal", "TriggerMode", "Sequence", "EventNumber", "AmbientTemperature", "UserLabel", "SerialNumber")
-  }else if(trigger.info=="Ultrafire_Video"){
-    Tag <- c("CreateDate")
+  # Check file types and define metadata tags for particular camera models
+  if(file.type %in% c(".jpg", ".JPG", "image")){
+    if(file.type == "image"){
+      message("You specified a general images tag for file.type. This function will search for .jpg files.")
+      file.type <- ".jpg"
+    }else{
+      file.type <- file.type
+    }
+
+    if(isTRUE(is.null(trigger.info))){
+      Tag <- c("DateTimeOriginal")
+    }else if(trigger.info == "Hyperfire2"){
+      Tag <- c("DateTimeOriginal", "TriggerMode", "Sequence", "EventNumber", "AmbientTemperature", "UserLabel", "SerialNumber")
+    }else if(trigger.info == "Ultrafire_image"){
+      Tag <- c("DateTimeOriginal", "TriggerMode", "Sequence", "EventNumber", "AmbientTemperature", "UserLabel", "SerialNumber")
+    }else{
+      Tag <- c("DateTimeOriginal")
+      message("Additional info is not supported for your camera model. Only date-time information will be provided. \nFor image file types, choose one of c('Hyperfire2', 'Ultrafire_image').")
+    }
+  }else if(file.type %in% c(".mp4", ".MP4", "video")){
+    if(file.type=="video"){
+      message("You specified a general video tag for file.type. This function will search for .mp4 files.")
+      file.type <- ".mp4"
+    }else{
+      file.type <- file.type
+    }
+
+    if(isTRUE(is.null(trigger.info))){
+      Tag <- c("CreateDate")
+    }else if(trigger.info == "Ultrafire_video"){
+      Tag <- c("CreateDate")
+    }else{
+      Tag <- c("CreateDate")
+      message("Additional info is not supported for your camera model. Only date-time information will be provided. \nFor video file types, choose one of c('Ultrafire_Video').")
+    }
   }else{
-    Tag <- c("DateTimeOriginal")
-    warning("Additional info is not supported for your camera model. Only date-time information will be provided. If you want additional information, choose one of c('Hyperfire2', 'Ultrafire_Video'). To figure out which metadata tags you want, run exiftoolr::exif_read('image file path').")
+    stop("You must specify an appropriate file.type. Choose one of c('.jpg', '.mp4', 'image', 'video')")
   }
 
   # Check if exiftool is installed within R
@@ -369,15 +445,6 @@ cameraRename3 <- function(in.dir, out.dir=NULL, file.type, trigger.info=NULL, re
     if(!all(dir.exists(unique(out.dir)))){
       stop("Some of your out.dirs do not exist. Do you need to create them?")
     }
-  }
-
-  # Check file types and ensure correct file types are used
-  if(file.type=="image"){
-    file.type <- ".jpg"
-  }else if(file.type=="video"){
-    file.type <- ".mp4"
-  }else{
-    file.type <- file.type
   }
 
   # Find files and check directory structure