+++ NEWS.md
@@ -20,3 +20,9 @@
 * Fixed an error with the description file
 * Fixed an error in the documentation for the timeConvert function where the lubridate package was improperly specified
 * Actually checked if the package could be built and fixed associated errors
+
+## Versopm 0.0.0.4 (2022-08-31)
+
+* Added an additional error code to the APFun_env function to ensure that camera names are included in the envdata file
+* Modified several functions so that the input requirements are more consistent. "timelapse" is used any time a timelapse file is required as the input. "x" is used for other inputs related to camera data. "ds" is used for other inputs not related to camera data.
+